requirements.txt
certifi==2020.11.8
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
future==0.18.2
h5py==3.1.0
imageio==2.9.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
olefile==0.46
opencv-python==4.5.1.48
openslide-python==1.1.2
pandas==1.2.0
Pillow @ file:///tmp/build/80754af9/pillow_1603822255246/work
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
threadpoolctl==2.1.0
tifffile==2021.1.11
torch==1.7.0
torchaudio==0.7.0a0+ac17b64
torchvision==0.8.1
tqdm==4.56.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
